Library/requirements.txt
Flask==3.0.0
pywebview==5.1